* Moved Plugin Reloading to the `AppBootstrapper` by way of Events 	* New `RendererInstanceEvents` 	* `RendererInstance` now extends `EventEmitter` * Converted Emitter access to an Accessor Function. * `SettingsOptions` UI/UX handling is now moved into   `SettingsRendererHelper`
