+++ src/scripts/types/Renderers.ts
@@ -4,10 +4,19 @@
  * @module
  */
 
+import { EventEmitter } from 'events';
 import { FormValidatorResults, ParsedJsonResults } from '../utils/Forms.js';
 import { TemplateMap } from '../utils/Templating.js';
 import { DisplayAccessor } from './Managers.js';
 import { PluginInstances, PluginSettingsBase } from './Plugin.js';
+
+/**
+ * Events that the {@link SettingsManager | `SettingsManager`} Emits.
+ */
+export enum RendererInstanceEvents {
+  /** Fired when Plugin List has been modified */
+  PLUGINS_STALE = 'rendererinstance::plugins-stale'
+}
 
 /**
  * Options for initializing the {@link RendererInstance | `RendererInstance`}.
@@ -17,7 +26,7 @@ import { PluginInstances, PluginSettingsBase } from './Plugin.js';
 export type RendererInstanceOptions<PluginSettings extends PluginSettingsBase> = {
   /** Accessor Function to get the Parsed JSON Results of processing a {@link FormEntry | `FormEntry[]`}. */
   getParsedJsonResults?: () => ParsedJsonResults | undefined;
-  /** Action Function to initiate Validating Settings for all Plugins */
+  /** Action Function to initiate Validating Settings for all Plugins. */
   validateSettings: () => FormValidatorResults<PluginSettings>;
   /** Accessor Function to retrieve the Settings masked. */
   getMaskedSettings: () => PluginSettings;
@@ -26,27 +35,25 @@ export type RendererInstanceOptions<PluginSettings extends PluginSettingsBase> =
    * @typeParam TemplateIDs - Union Type of accepted `TemplateIDs`.
    */
   getTemplates: <TemplateIDs extends string>() => TemplateMap<TemplateIDs>;
-  /** Accessor Function for Settings */
+  /** Accessor Function for Settings. */
   getSettings: () => PluginSettings;
   /**
-   * Action Function to Set Settings
+   * Action Function to Set Settings.
    *
    * @param settings - Settings to store for the System.
    * @param forceEncode - Whether or not to force encoding appropriate values.
    */
   setSettings: (settings: PluginSettings, forceEncode?: boolean) => void;
-  /** Accessor Function for Registered Plugins */
+  /** Accessor Function for Registered Plugins. */
   getPlugins: () => PluginInstances<PluginSettings>;
-  /** Action Function to Reinitialize Plugins */
-  pluginLoader: () => void;
-  /** Accessor for Display Manager */
+  /** Accessor for Display Manager. */
   display: DisplayAccessor;
 };
 
 /**
  * Instance Typing of a Renderer.
  */
-export type RendererInstance = {
+export type RendererInstance = EventEmitter & {
   init(): Promise<void>;
 };
